+++ packages/safe-fn/src/types/catch-handler.ts
